@@ -1,14 +1,20 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
+
+/**
+ *  Shift
+ *  Copyright (c) Vinh Nguyen 2021
+ *  MIT license, see LICENSE file for details
+ */
 
 import PackageDescription
 
 let package = Package(
     name: "Shift",
     platforms: [
-        .iOS(.v14),
-        .macOS(.v10_15),
-        .watchOS(.v6)
+        .iOS(.v15),
+        .macOS(.v12),
+        .watchOS(.v8)
     ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
